+++ src/Braintree/PaymentMethodNonceDetails.cs
@@ -10,6 +10,7 @@ namespace Braintree
         public virtual string CardType { get; protected set; }
         public virtual string Username { get; protected set; }
         public virtual string VenmoUserId { get; protected set; }
+        public virtual PaymentMethodNonceDetailsPayerInfo PayerInfo { get; protected set; }
 
         protected internal PaymentMethodNonceDetails(NodeWrapper node)
         {
@@ -19,6 +20,11 @@ namespace Braintree
             LastFour = node.GetString("last-four");
             Username = node.GetString("username");
             VenmoUserId = node.GetString("venmo-user-id");
+            var payerInfoNode = node.GetNode("payer-info");
+            if (payerInfoNode != null)
+            {
+                PayerInfo = new PaymentMethodNonceDetailsPayerInfo(payerInfoNode);
+            }
         }
 
         [Obsolete("Mock Use Only")]