+++ src/Braintree/Properties/AssemblyInfo.cs
@@ -15,6 +15,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyCulture("")]
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("Braintree.Tests")]
 [assembly: CLSCompliant(true)]
+//[assembly: AssemblyKeyFileAttribute("../../braintreeSgKey.snk")]
 
 // Setting ComVisible to false makes the types in this assembly not visible
 // to COM components.  If you need to access a type in this assembly from
@@ -31,5 +32,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("4.12.0.0")]
-[assembly: AssemblyFileVersion("4.12.0.0")]
+[assembly: AssemblyVersion("4.13.0.0")]
+[assembly: AssemblyFileVersion("4.13.0.0")]