+++ src/Braintree/TransactionThreeDSecurePassThruRequest.cs
@@ -7,6 +7,11 @@ namespace Braintree
         public string EciFlag { get; set; }
         public string Cavv { get; set; }
         public string Xid { get; set; }
+        public string AuthenticationResponse { get; set; }
+        public string DirectoryResponse { get; set; }
+        public string CavvAlgorithm { get; set; }
+        public string DsTransactionId { get; set; }
+        public string ThreeDSecureVersion { get; set; }
 
         public override string ToXml(string root)
         {
@@ -23,7 +28,12 @@ namespace Braintree
             return new RequestBuilder(root).
                 AddElement("eci-flag", EciFlag).
                 AddElement("cavv", Cavv).
-                AddElement("xid", Xid);
+                AddElement("xid", Xid).
+                AddElement("three-d-secure-version", ThreeDSecureVersion).
+                AddElement("authentication_response", AuthenticationResponse).
+                AddElement("directory_response", DirectoryResponse).
+                AddElement("cavv_algorithm", CavvAlgorithm).
+                AddElement("ds_transaction_id", DsTransactionId);
         }
     }
 }