+++ test/Braintree.Tests.Integration/CreditCardIntegrationTest.cs
@@ -1,4 +1,5 @@
 using Braintree.Exceptions;
+using Braintree.Test;
 using Braintree.TestUtil;
 using NUnit.Framework;
 using System;
@@ -1482,6 +1483,35 @@ namespace Braintree.Tests.Integration
           };
           Result<CreditCard> result = gateway.CreditCard.Create(request);
           Assert.IsTrue(result.IsSuccess());
+        }
+
+        [Test]
+        public void CreateWithThreeDSecureNonce()
+        {
+          Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+          CreditCardRequest request = new CreditCardRequest
+          {
+            CustomerId = customer.Id,
+            PaymentMethodNonce = Nonce.ThreeDSecureVisaFullAuthentication,
+            Options = new CreditCardOptionsRequest()
+            {
+                VerifyCard = true
+            },
+          };
+          Result<CreditCard> result = gateway.CreditCard.Create(request);
+
+          CreditCard card = result.Target;
+
+          CreditCardVerification verification = card.Verification;
+
+          Assert.AreEqual("Y", verification.ThreeDSecureInfo.Enrolled);
+          Assert.AreEqual("cavv_value", verification.ThreeDSecureInfo.Cavv);
+          Assert.AreEqual("05", verification.ThreeDSecureInfo.EciFlag);
+          Assert.AreEqual("authenticate_successful", verification.ThreeDSecureInfo.Status);
+          Assert.AreEqual("1.0.2", verification.ThreeDSecureInfo.ThreeDSecureVersion);
+          Assert.AreEqual("xid_value", verification.ThreeDSecureInfo.Xid);
+          Assert.IsTrue(verification.ThreeDSecureInfo.LiabilityShifted);
+          Assert.IsTrue(verification.ThreeDSecureInfo.LiabilityShiftPossible);
         }
 
         [Test]