+++ test/Braintree.Tests.Integration/CreditCardVerificationIntegrationTest.cs
@@ -57,6 +57,42 @@ namespace Braintree.Tests.Integration
         }
 
         [Test]
+        public void Create_ReturnsSuccessfulResponseWithNetworkResponseCodeText()
+        {
+            var request = new CreditCardVerificationRequest
+            {
+                CreditCard = new CreditCardVerificationCreditCardRequest
+                {
+                    Number = SandboxValues.CreditCardNumber.VISA,
+                    ExpirationDate = "05/2009",
+                    BillingAddress = new CreditCardAddressRequest
+                    {
+                        CountryName = "Greece",
+                        CountryCodeAlpha2 = "GR",
+                        CountryCodeAlpha3 = "GRC",
+                        CountryCodeNumeric = "300"
+                    }
+                },
+                Options = new CreditCardVerificationOptionsRequest
+                {
+                    MerchantAccountId = MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID,
+                    Amount = "5.00"
+                }
+            };
+
+            Result<CreditCardVerification> result = gateway.CreditCardVerification.Create(request);
+            Assert.IsTrue(result.IsSuccess());
+            CreditCardVerification verification = result.Target;
+            Assert.AreEqual(verification.MerchantAccountId,
+                            MerchantAccountIDs.NON_DEFAULT_MERCHANT_ACCOUNT_ID);
+            Assert.AreEqual("1000", verification.ProcessorResponseCode);
+            Assert.AreEqual("Approved", verification.ProcessorResponseText);
+            Assert.AreEqual(ProcessorResponseType.APPROVED, verification.ProcessorResponseType);
+            Assert.AreEqual("XX", verification.NetworkResponseCode);
+            Assert.AreEqual("sample network response text", verification.NetworkResponseText);
+        }
+
+        [Test]
         public void Create_ReturnsUnsuccessfulResponse()
         {
             var request = new CreditCardVerificationRequest