+++ test/Braintree.Tests/PaymentMethodNonceTest.cs
@@ -51,6 +51,46 @@ namespace Braintree.Tests
         }
 
         [Test]
+        public void ParsesNodeCorrectlyWithPayPalDetails()
+        {
+            string xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<payment-method-nonce>" +
+                "  <type>PayPalAccount</type>" +
+                "  <nonce>fake-paypal-account-nonce</nonce>" +
+                "  <description></description>" +
+                "  <consumed type=\"boolean\">false</consumed>" +
+                "  <details>" +
+                "    <payer-info>" +
+                "      <email>jane.doe@paypal.com</email>" +
+                "      <first-name>first</first-name>" +
+                "      <last-name>last</last-name>" +
+                "      <payer-id>pay-123</payer-id>" +
+                "      <country-code>US</country-code>" +
+                "    </payer-info>" +
+                "  </details>" +
+                "</payment-method-nonce>";
+
+            XmlDocument doc = new XmlDocument();
+            doc.LoadXml(xml);
+            XmlNode newNode = doc.DocumentElement;
+
+            var node = new NodeWrapper(newNode);
+
+            var result = new ResultImpl<PaymentMethodNonce>(node, gateway);
+
+            Assert.IsNotNull(result.Target);
+            Assert.AreEqual("fake-paypal-account-nonce", result.Target.Nonce);
+            Assert.AreEqual("PayPalAccount", result.Target.Type);
+            Assert.IsNotNull(result.Target.Details);
+            Assert.IsNotNull(result.Target.Details.PayerInfo);
+            Assert.AreEqual("jane.doe@paypal.com", result.Target.Details.PayerInfo.Email);
+            Assert.AreEqual("first", result.Target.Details.PayerInfo.FirstName);
+            Assert.AreEqual("last", result.Target.Details.PayerInfo.LastName);
+            Assert.AreEqual("pay-123", result.Target.Details.PayerInfo.PayerId);
+            Assert.AreEqual("US", result.Target.Details.PayerInfo.CountryCode);
+        }
+
+        [Test]
         public void ParsesNodeCorrectlyWithVenmoDetails()
         {
             string xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
@@ -158,6 +198,49 @@ namespace Braintree.Tests
             Assert.AreEqual("CreditCard", result.Target.Type);
             Assert.IsNull(result.Target.ThreeDSecureInfo);
             Assert.IsNull(result.Target.Details);
+        }
+
+        [Test]
+        public void ParseThreeDSecureInfo()
+        {
+            string xml = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>" +
+                "<payment-method-nonce>" +
+                "  <type>CreditCard</type>" +
+                "  <nonce>fake-valid-nonce</nonce>" +
+                "  <description>ending in 22</description>" +
+                "  <consumed type=\"boolean\">false</consumed>" +
+                "  <three-d-secure-info>" +
+                "    <enrolled>Y</enrolled>" +
+                "    <status>authenticate_successful</status>" +
+                "    <liability-shifted>true</liability-shifted>" +
+                "    <liability-shift-possible>true</liability-shift-possible>" +
+                "    <cavv>cavv-value</cavv>" +
+                "    <xid>xid-value</xid>" +
+                "    <ds-transaction-id>ds-trx-id-value</ds-transaction-id>" +
+                "    <eci-flag>06</eci-flag>" +
+                "    <three-d-secure-version>2.0.1</three-d-secure-version>" +
+                "  </three-d-secure-info>" +
+                "  <details nil=\"true\"/>" +
+                "</payment-method-nonce>";
+
+            XmlDocument doc = new XmlDocument();
+            doc.LoadXml(xml);
+            XmlNode newNode = doc.DocumentElement;
+
+            var node = new NodeWrapper(newNode);
+
+            var result = new ResultImpl<PaymentMethodNonce>(node, gateway);
+            var threeDSecureInfo = result.Target.ThreeDSecureInfo;
+            Assert.IsNotNull(result.Target.ThreeDSecureInfo);
+            Assert.AreEqual("Y", threeDSecureInfo.Enrolled);
+            Assert.AreEqual("authenticate_successful", threeDSecureInfo.Status);
+            Assert.IsTrue(threeDSecureInfo.LiabilityShifted);
+            Assert.IsTrue(threeDSecureInfo.LiabilityShiftPossible);
+            Assert.AreEqual("cavv-value", threeDSecureInfo.Cavv);
+            Assert.AreEqual("xid-value", threeDSecureInfo.Xid);
+            Assert.AreEqual("ds-trx-id-value", threeDSecureInfo.DsTransactionId);
+            Assert.AreEqual("06", threeDSecureInfo.EciFlag);
+            Assert.AreEqual("2.0.1", threeDSecureInfo.ThreeDSecureVersion);
         }
     }
 }